#23　APIManagerを作り直しました。　めんどくさかったのでこのコミットでいろいろやっちゃった
・Get,PostAPIManagerを削除
・User,Tweet,Auth,SearchAPIManagerを代わりに作成
・それに合わせて、loginViewController,TimeLineViewControllerを修正
