+++ Tsuitta/Model/API/APIClient.swift
@@ -2,14 +2,14 @@ import TwitterKit
 
 class APIClient {
     
-    private let baseURL = "https://api.twitter.com"
-    private let version = "/1.1"
+    static private let baseURL = "https://api.twitter.com"
+    static private let version = "/1.1"
     
-    init() {
+    private init() {
         
     }
     
-    private func sendRequest(method: String, path: String, parameter: [NSObject : AnyObject]? = nil,  callback: TWTRNetworkCompletion){
+    static private func sendRequest(method: String, path: String, parameter: [NSObject : AnyObject]? = nil,  callback: TWTRNetworkCompletion){
         //TODO: clientError の使い方よくわからないよ！
         var clientError: NSError?
         let endpoint = baseURL + version + path
@@ -21,11 +21,11 @@ class APIClient {
         })
     }
     
-    func get(path: String, parameter: [NSObject : AnyObject]? = nil,  callback: TWTRNetworkCompletion) {
+    static func get(path: String, parameter: [NSObject : AnyObject]? = nil,  callback: TWTRNetworkCompletion) {
         sendRequest("GET", path: path, parameter: parameter, callback: callback)
     }
     
-    func post(path: String, parameter: [NSObject : AnyObject]? = nil,  callback: TWTRNetworkCompletion) {
+    static func post(path: String, parameter: [NSObject : AnyObject]? = nil,  callback: TWTRNetworkCompletion) {
         sendRequest("POST", path: path, parameter: parameter, callback: callback)
     }
 }
