--- Tsuitta/Model/API/PostAPIManager.swift
@@ -1,7 +1,0 @@
-import TwitterKit
-import SwiftyJSON
-
-/// Post系のAPIメソッドがまとまっているクラスです
-class PostAPIManager {
-    
-}
